requirements-dev.txt
-r requirements.txt

# Development tools
aws-cdk-lib>=2.0.0
constructs>=10.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
pytest>=7.0.0
pytest-django>=4.5.0 